% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_rhist.R
\name{collect_rhist}
\alias{collect_rhist}
\alias{collect_session_packages}
\title{Collect command and session history}
\usage{
collect_rhist(rhist_path = installed_rhist_path())

collect_session_packages(rhist_path = installed_rhist_path())
}
\arguments{
\item{rhist_path}{Path of rhist database}
}
\value{
A data frame
}
\description{
A data frame with command text and session info
}
\section{Functions}{
\itemize{
\item \code{collect_session_packages}: A data frame with package info for each session
}}

