% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_rhist.R
\name{default_rhist_path}
\alias{default_rhist_path}
\title{Create a default path for db within the user's home directory}
\usage{
default_rhist_path()
}
\description{
Create a default path for db within the user's home directory
}
