% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_rhist.R
\name{install_rhist}
\alias{install_rhist}
\title{Install rhist}
\usage{
install_rhist(dbpath = default_rhist_path())
}
\arguments{
\item{dbpath}{Character. The path where an SQLite database should be
installed. Defaults to the user's home directory.}
}
\description{
Initializes a new .rhistory.sqlite db and provides code to add a .Last
function to your .Rprofile
}
