% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhist.R
\name{save_session_history}
\alias{save_session_history}
\title{Save current session history to specified SQLite database}
\usage{
save_session_history(dbpath = installed_rhist_path())
}
\arguments{
\item{dbpath}{Path to rhist database.}
}
\description{
Save current session history to specified SQLite database
}
